tp/61028 - Ponce, Rodrigo Exequíel/tp6/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.5",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.27",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.32.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
